Move interfaces to seperate folder
